try BigIntegerField Postgres fix - 2

+++ remenis/core/models.py
@@ -28,7 +28,7 @@ class BigIntegerField(fields.IntegerField):
             
 class User(models.Model):
     fbid = models.CharField(max_length=20)
-    facebook_id = models.BigIntegerField(default=0, primary_key=True)
+    facebook_id = models.IntegerField(default=0, primary_key=True)
     first_name = models.CharField(max_length=50, blank=True)
     last_name = models.CharField(max_length=50, blank=True)
     full_name = models.CharField(max_length=100)
